npm run build first builds js-xdr, declare-constants script can now run multiple times
