+++ .gitignore
@@ -14,3 +14,4 @@
 **/yarn.lock
 **/package-lock.json
 **/*.yarn-integrity
+**/dist/
