+++ autogenerator/js-xdr/package.json
@@ -4,7 +4,9 @@
   "description": "",
   "main": "dist/src/index.js",
   "types": "dist/types/src/index.d.ts",
-  "scripts": {},
+  "scripts": {
+    "build": "tsc"
+  },
   "author": "SatoshiPay Ltd.",
   "license": "Apache-2.0",
   "devDependencies": {
